feat: rename and restructure ErrorMessageProps type
