+++ app/components/formElements/RadioGroup.tsx
@@ -1,6 +1,6 @@
 import { useField } from "@rvf/react-router";
 import { type ReactNode, useState } from "react";
-import { type ErrorMessageProps } from ".";
+import { type ErrorMessageProps } from "../types";
 import InputError from "./InputError";
 import Radio from "./Radio";
 import { useJsAvailable } from "../hooks/useJsAvailable";
