+++ app/components/formElements/Select.tsx
@@ -1,7 +1,7 @@
 import { useField } from "@rvf/react-router";
 import classNames from "classnames";
 import type { ReactNode } from "react";
-import { type ErrorMessageProps } from ".";
+import { type ErrorMessageProps } from "../types";
 import InputError from "./InputError";
 import InputLabel from "./InputLabel";
 import { widthClassname } from "../width";
