+++ app/routes/beratungshilfe.zustaendiges-gericht.auswahl.$PLZ.tsx
@@ -11,12 +11,12 @@ import Button from "~/components/Button";
 import ButtonContainer from "~/components/ButtonContainer";
 import Container from "~/components/Container";
 import CourtFinderHeader from "~/components/CourtFinderHeader";
-import { type ErrorMessageProps } from "~/components/formElements";
 import AutoSuggestInput from "~/components/formElements/AutoSuggestInput";
 import Input from "~/components/formElements/Input";
 import Heading from "~/components/Heading";
 import { ReportProblem } from "~/components/reportProblem/ReportProblem";
 import RichText from "~/components/RichText";
+import { type ErrorMessageProps } from "~/components/types";
 import { streetHouseNumberSchema } from "~/domains/shared/formular/persoenlicheDaten/userData";
 import { fetchMeta, fetchTranslations } from "~/services/cms/index.server";
 import { edgeCaseStreets } from "~/services/gerichtsfinder/amtsgerichtData.server";
